Hook into Moo::Role via role_application_steps rather than using an after modifier on apply_roles_to_package.
Firing hooks after apply_roles_to_package is logical, but if a role is
consumed via Moo::Role::with, the after modifier is put in place too
late:

1. Moo::Role::with calls apply_roles_to_package
2. apply_roles_to_package loads the role package
3. the role package applies the after modifier to apply_roles_to_package

However, by by step 3, execution is already in the *unmodified*
apply_roles_to_package, so the hook never gets fired.

Internally, `apply_roles_to_package` does things in this order:

1. load the roles
2. apply the roles to the package
3. call `role_application_steps` to get the list of things to do
4. do the things to do

Conveniently, role_application_steps is called after the role is
loaded, so the role can modify it before execution gets to that part
of the algorithm, providing an opportunity for the role to
install a hook while `apply_roles_to_package` is executing, allowing,
essentially, the chicken to come before the egg, or (vice versa if
you're a bigger fan of eggs).

This commit modifies role_application_steps, adding a final step
which will fire the hooks.

The drawback of this approach is that it relies upon a detail of
the implementation. There's no technical reason that `apply_roles_to_package`
couldn't get the list of steps from `role_application_steps` before it
loads the roles.  At this point I doubt that that code would change,
but there's no guarantee.

From a caller's perspective the only difference between this approach
and the previous one is that (at least as of Role::Tiny 2.002004), the
final registration of the role application (e.g. recording it in
%Role::Tiny::APPLIED_TO) would done *after* the hook is fired.

I don't see another place to hook into the role application process
that would solve the problem.  FWIW, executing the following code will
also do the trick, if done before any of the roles are loaded,

  require MooX::CaptainHook;
  'Moo::Role'->apply_single_role_to_package('Moo::Role', MooX::CaptainHook::ON_APPLICATION())
          unless Role::Tiny::does_role('Moo::Role', MooX::CaptainHook::ON_APPLICATION());

but that requires that the client code using the roles know that the
roles knows they are using `MooX::CaptainHook`, which is TMI.