long class names no longer cause "Identifier too long" error
Perl allows only so many characters in an identfier, and composition
with Moo roles and classes can cause a package name's length to
approach the legal limit.

Moo::Role ensures that the package name itself remains legal, but a
fully qualified variable name may be too long.

To work around this, expressions involving a fully qualified variable
name are rewritten into a 'do' block which enters the package, and
then uses the relative variable name.